Renames:
EntityOverview -> ResourceMetricOverview

HealthPane -> ResourceHealthOverview
